Config var env

+++ api-suggested-value/src/main/java/pa/com/sura/apisuggestedvalue/models/dto/SuggestedValueDto.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class SuggestedValueDto implements Serializable{
   private String currentYear;
-  private Long suggestedValue;
+  private double suggestedValue; 
 }
